+++ e2e/utilities/consts.ts
@@ -1,2 +1,3 @@
 export const TESTS_OUTPUT_FOLDER = 'tests-output';
 export const SCREENSHOTS_FOLDER = 'tests-output/screenshots';
+export const COOLSTORE_REPO_URL = 'https://github.com/konveyor-ecosystem/coolstore';
